% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7icons-dependencies.R
\name{add_f7icons_deps}
\alias{add_f7icons_deps}
\title{Framework7 icon dependencies}
\usage{
add_f7icons_deps(tag)
}
\arguments{
\item{tag}{Element to attach the dependencies.}
}
\description{
This function attaches icon dependencies to the given tag.
}
\keyword{internal}
