% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/framework7-dependencies.R
\name{add_framework7_deps}
\alias{add_framework7_deps}
\title{framework7 dependencies utils}
\usage{
add_framework7_deps(tag)
}
\arguments{
\item{tag}{Element to attach the dependencies.}
}
\description{
This function attaches framework7. dependencies to the given tag
}
