% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwa-dependencies.R
\name{add_pwa_deps}
\alias{add_pwa_deps}
\title{PWA dependencies utils}
\usage{
add_pwa_deps(tag)
}
\arguments{
\item{tag}{Element to attach the dependencies.}
}
\description{
This function attaches PWA manifest and icons to the given tag
}
