% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwacompat-dependencies.R
\name{add_pwacompat_deps}
\alias{add_pwacompat_deps}
\title{pwacompat dependencies utils}
\usage{
add_pwacompat_deps(tag)
}
\arguments{
\item{tag}{Element to attach the dependencies.}
}
\description{
This function attaches pwacompat. dependencies to the given tag
}
