% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyMobile-dependencies.R
\name{add_shinyMobile_deps}
\alias{add_shinyMobile_deps}
\title{shinyMobile dependencies utils}
\usage{
add_shinyMobile_deps(tag)
}
\arguments{
\item{tag}{Element to attach the dependencies.}
}
\description{
This function attaches shinyMobile dependencies to the given tag
}
