% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7-tabs.R
\name{f7Back}
\alias{f7Back}
\alias{f7Next}
\title{Framework7 back button}
\usage{
f7Back(targetId)

f7Next(targetId)
}
\arguments{
\item{targetId}{\link{f7Tabs} id.}
}
\description{
\link{f7Back} is a button to go back in \link{f7Tabs}.

\link{f7Next} is a button to go next in \link{f7Tabs}.
}
