% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7-inputs.R
\name{buildConfig}
\alias{buildConfig}
\title{Build config tag for JavaScript}
\usage{
buildConfig(id, config)
}
\description{
See \url{https://unleash-shiny.rinterface.com/shiny-input-system#boxes-on-steroids-more}
}
\keyword{internal}
