% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7-inputs.R
\name{buildPickerInput}
\alias{buildPickerInput}
\title{Build input tag for picker elements}
\usage{
buildPickerInput(id, label, config, class, placeholder = NULL)
}
\description{
Build input tag for picker elements
}
\keyword{internal}
