% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_app_link}
\alias{create_app_link}
\title{Create shinylive iframe}
\usage{
create_app_link(app_code, mode = c("app", "editor"), header = TRUE)
}
\arguments{
\item{app_code}{base64 app code. You can create it from https://shinylive.io/r
by writing code and click on share and copy the link. The code is located at
the end of the url.}

\item{mode}{How to display the shinylive app. Default to app mode.}

\item{header}{Whether to display the shinylive header. Default to TRUE.}
}
\description{
Create shinylive iframe
}
\keyword{internal}
