% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview_mobile.R
\name{create_app_ui}
\alias{create_app_ui}
\title{Create the app UI}
\usage{
create_app_ui(iframe, device, color, landscape, static = FALSE)
}
\arguments{
\item{iframe}{iframe tag designed by \link{preview_mobile}.}

\item{device}{See \link{preview_mobile} input.}

\item{color}{See \link{preview_mobile} input.}

\item{landscape}{See \link{preview_mobile} input.}
}
\description{
Create the app UI
}
\keyword{Internal}
