% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7-inputs.R
\name{createInputLayout}
\alias{createInputLayout}
\title{Create common input layout}
\usage{
createInputLayout(
  ...,
  label = NULL,
  media = NULL,
  floating = FALSE,
  outline = FALSE,
  info = NULL,
  dropdown = FALSE,
  clear = TRUE
)
}
\description{
See \url{https://framework7.io/docs/inputs#inputs-layout}.
}
\keyword{internal}
