% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7-inputs.R
\name{createInputTag}
\alias{createInputTag}
\title{Create an input tag}
\usage{
createInputTag(inputId, value = NULL, type, placeholder, ...)
}
\description{
Useful for text inputs, password input ...
}
\keyword{internal}
