% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7-inputs.R
\name{createOptions}
\alias{createOptions}
\title{Generates a list of option}
\usage{
createOptions(
  inputId,
  choices,
  selected,
  position = "left",
  has_media_list = FALSE,
  type
)
}
\arguments{
\item{inputId}{Radio input id.}

\item{choices}{List of choices.}

\item{selected}{Selected value}

\item{position}{Check mark position.}

\item{has_media_list}{For custom choices.}

\item{type}{Choose either "checkbox" or "radio"}
}
\description{
For \link{f7Radio} and \link{f7CheckboxGroup}
}
\keyword{internal}
