% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7-inputs.R
\name{createRadioOptions}
\alias{createRadioOptions}
\title{Generates a list of option for \link{f7Radio}}
\usage{
createRadioOptions(
  choices,
  selected,
  inputId,
  position = "left",
  has_media_list = FALSE
)
}
\arguments{
\item{choices}{List of choices.}

\item{selected}{Selected value}

\item{inputId}{Radio input id.}

\item{position}{Check mark position.}

\item{has_media_list}{For custom choices.}
}
\description{
Generates a list of option for \link{f7Radio}
}
\keyword{internal}
