% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7-inputs.R
\name{createSelectOptions}
\alias{createSelectOptions}
\title{Create option html tag based on choice input}
\usage{
createSelectOptions(choices, selected)
}
\arguments{
\item{choices}{Vector of possibilities.}

\item{selected}{Default selected value.}
}
\description{
Used by \link{f7SmartSelect} and \link{f7Select}
}
