% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7-inputs.R, R/f7-tabs.R
\name{f7checkBox}
\alias{f7checkBox}
\alias{f7checkBoxGroup}
\alias{f7InsertTab}
\alias{f7RemoveTab}
\title{Framework7 checkbox}
\usage{
f7checkBox(inputId, label, value = FALSE)

f7checkBoxGroup(inputId, label, choices = NULL, selected = NULL)

f7InsertTab(
  id,
  tab,
  target,
  position = c("before", "after"),
  select = FALSE,
  session = shiny::getDefaultReactiveDomain()
)

f7RemoveTab(id, target, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{id}{\link{f7Tabs} id.}

\item{tab}{\link{f7Tab} to insert.}

\item{target}{\link{f7Tab} after of before which the new tab will be inserted.}

\item{position}{Insert before or after: \code{c("before", "after")}.}

\item{select}{Whether to select the newly inserted tab. FALSE by default.}

\item{session}{Shiny session object.}
}
\description{
\link{f7checkBox} creates a checkbox input (use \link{f7Checkbox}).

Creates a checkbox group input (use )

Inserts a \link{f7Tab} in a \link{f7Tabs} (use \link{f7InsertTab}).

Removes a \link{f7Tab} in a \link{f7Tabs} (use \link{removeF7Tab}).
}
