% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7Accordion.R
\name{f7AccordionItem}
\alias{f7AccordionItem}
\title{Create a Framework7 accordion item}
\usage{
f7AccordionItem(..., title = NULL, open = FALSE)
}
\arguments{
\item{...}{Item content.}

\item{title}{Item title.}

\item{open}{Whether the item is open at start. FALSE by default.}
}
\description{
Build a Framework7 accordion item
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
