% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7Messages.R
\name{f7AddMessages}
\alias{f7AddMessages}
\title{Update \link{f7Messages} on the server side.}
\usage{
f7AddMessages(
  id,
  messages,
  showTyping = FALSE,
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{id}{Reference to link{f7Messages} container.}

\item{messages}{List of \link{f7Messages}.}

\item{showTyping}{Show typing when a new message comes. Default to FALSE.}

\item{session}{SHiny session object}
}
\description{
Update \link{f7Messages} on the server side.
}
