% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typography.R
\name{f7Align}
\alias{f7Align}
\title{Framework7 align utility}
\usage{
f7Align(tag, side = c("left", "center", "right", "justify"))
}
\arguments{
\item{tag}{Tag to align.}

\item{side}{Side to align: "left", "center", "right" or "justify".}
}
\description{
\link{f7Align} is an alignment utility for items.
}
\examples{
if(interactive()){
 library(shiny)
 library(shinyMobile)

 shinyApp(
   ui = f7Page(
    title = "Align",
    f7SingleLayout(
     navbar = f7Navbar(title = "f7Align"),
     f7Row(
      f7Align(h1("Left"), side = "left"),
      f7Align(h1("Center"), side = "center"),
      f7Align(h1("Right"), side = "right")
     )
    )
   ),
   server = function(input, output) {}
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
