% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7-inputs.R
\name{f7AutoComplete}
\alias{f7AutoComplete}
\title{Create a Framework7 autocomplete input}
\usage{
f7AutoComplete(inputId, label, placeholder = NULL, value = choices[1],
  choices, typeahead = TRUE, expandInput = TRUE, type = c("popup",
  "page", "dropdown"), dropdownPlaceholderText = NULL)
}
\arguments{
\item{inputId}{Autocomplete input id.}

\item{label}{Autocomplete label.}

\item{placeholder}{Text to write in the container.}

\item{value}{Autocomplete initial value, if any.}

\item{choices}{Autocomplete choices.}

\item{typeahead}{Enables type ahead, will prefill input
value with first item in match.}

\item{expandInput}{If TRUE then input which is used as
item-input in List View will be expanded to full
screen wide during dropdown visible.}

\item{type}{Defines how to open Autocomplete,
can be page or popup (for Standalone) or dropdown.}

\item{dropdownPlaceholderText}{Specify dropdown placeholder text.}
}
\description{
Build a Framework7 autocomplete input
}
\examples{
if(interactive()){
 library(shiny)
 library(shinyF7)

 shinyApp(
   ui = f7Page(
    title = "My app",
    f7SingleLayout(
     navbar = f7Navbar(title = "f7Picker"),
     f7AutoComplete(
      inputId = "myautocomplete",
      placeholder = "Some text here!",
      dropdownPlaceholderText = "Try to type Apple",
      label = "Type a fruit name",
      type = "dropdown",
      choices = c('Apple', 'Apricot', 'Avocado', 'Banana', 'Melon',
       'Orange', 'Peach', 'Pear', 'Pineapple')
     ),
     textOutput("autocompleteval")
    )
   ),
   server = function(input, output) {
    output$autocompleteval <- renderText(input$myautocomplete)
   }
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
