% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7Appbar.R
\name{f7Back}
\alias{f7Back}
\title{Create a framework 7 back button}
\usage{
f7Back(targetId)
}
\arguments{
\item{targetId}{\link{f7Tabs} id.}
}
\description{
This buttons allows to switch between multiple \link{f7Tab}.
}
