% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7Block.R
\name{f7BlockFooter}
\alias{f7BlockFooter}
\title{Create a Framework7 block footer}
\usage{
f7BlockFooter(text = NULL)
}
\arguments{
\item{text}{Footer text.}
}
\description{
Build a Framework7 block footer
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
