% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7Block.R
\name{f7BlockHeader}
\alias{f7BlockHeader}
\title{Create a Framework7 block header}
\usage{
f7BlockHeader(text = NULL)
}
\arguments{
\item{text}{Header text.}
}
\description{
Build a Framework7 block header
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
