% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7Button.R
\name{f7Button}
\alias{f7Button}
\title{Create a Framework7 button}
\usage{
f7Button(inputId = NULL, label = NULL, src = NULL, color = NULL,
  fill = FALSE, outline = FALSE, shadow = FALSE, rounded = FALSE,
  size = NULL)
}
\arguments{
\item{inputId}{The input slot that will be used to access the value.}

\item{label}{The contents of the button or link–usually a text label,
but you could also use any other HTML, like an image.}

\item{src}{Button link.}

\item{color}{Button color. See here for valid colors \url{https://framework7.io/docs/badge.html}.}

\item{fill}{Fill style. FALSE by default.}

\item{outline}{Outline style. FALSE by default.}

\item{shadow}{Button shadow. FALSE by default. Only for material design.}

\item{rounded}{Round style. FALSE by default.}

\item{size}{Button size. NULL by default but also "large" or "small".}
}
\description{
Build a Framework7 button
}
\examples{
if(interactive()){
 library(shiny)
 library(shinyF7)

 shiny::shinyApp(
  ui = f7Page(
    title = "Buttons",
    f7Init(theme = "md"),
    f7Button(color = "blue", label = "My button", src = "https://www.google.com"),
    f7Button(color = "red", label = "Action Button", inputId = "button2", fill = TRUE),
    f7Button(color = "green", label = "My button", outline = TRUE),
    f7Button(color = "yellow", label = "My button", rounded = TRUE),
    f7Button(color = "pink", label = "My button", shadow = TRUE),
    f7Button(color = "purple", label = "My button", size = "large"),
    f7Button(color = "orange", label = "My button", size = "small"),

    verbatimTextOutput("val")
  ),
  server = function(input, output) {
   output$val <- renderPrint(input$button2)
  }
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
