% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7-inputs.R
\name{f7checkboxInput}
\alias{f7checkboxInput}
\title{Create a F7 Checkbox}
\usage{
f7checkboxInput(id, label, value = FALSE, width = NULL)
}
\arguments{
\item{id}{The input slot that will be used to access the value.}

\item{label}{Display label for the control, or \code{NULL} for no label.}

\item{value}{Initial value (\code{TRUE} or \code{FALSE}).}

\item{width}{The width of the input, e.g. \code{'400px'}, or \code{'100\%'};
see \code{\link{validateCssUnit}}.}
}
\description{
Create a F7 Checkbox
}
\examples{
if(interactive()){
 library(shiny)
 library(shinyF7)

 shiny::shinyApp(
   ui = f7Page(
    title = "My app",
    f7Card(
     f7checkboxInput(
      id = "check",
      label = "Checkbox",
      value = FALSE
     ),
     verbatimTextOutput("test")
    )
   ),
   server = function(input, output) {
    output$test <- renderPrint({input$check})
   }
 )
}
}
