% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7Grid.R
\name{f7Col}
\alias{f7Col}
\title{Create a Framework7 column container}
\usage{
f7Col(..., width = NULL, resizable = FALSE)
}
\arguments{
\item{...}{Column content. The width is automatically handled depending
on the number of columns.}

\item{width}{Width of column, an integer ranging from 0 to 100 giving the width
in percentage.}

\item{resizable}{Whether to create a resizable column.}
}
\description{
Build a Framework7 column container
}
\note{
The dark theme does not work for items embedded in a column. Use \link{f7Flex}
instead.
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
