% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7-inputs.R
\name{f7Date}
\alias{f7Date}
\title{Create a Framework7 date input}
\usage{
f7Date(inputId, label, value = NULL, min = NULL, max = NULL,
  format = "yyyy-mm-dd")
}
\arguments{
\item{inputId}{Date input id.}

\item{label}{Input label.}

\item{value}{Start value.}

\item{min}{Minimum date.}

\item{max}{Maximum date.}

\item{format}{Date format: "yyyy-mm-dd", for instance.}
}
\description{
Create a Framework7 date input
}
