% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7Page.R
\name{f7DefaultOptions}
\alias{f7DefaultOptions}
\title{shinyMobile app default options}
\usage{
f7DefaultOptions()
}
\value{
A list of options to pass in
\link{f7Page}.
}
\description{
List of default custom options.
}
