% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7Card.R
\name{f7ExpandableCard}
\alias{f7ExpandableCard}
\title{Create a Framework7 expandable card}
\usage{
f7ExpandableCard(..., title = NULL, subtitle = NULL,
  titleColor = c("white", "black"), color = NULL)
}
\arguments{
\item{...}{Card content.}

\item{title}{Card title.}

\item{subtitle}{Card subtitle.}

\item{titleColor}{Title text color. Choose among "white" or "black".}

\item{color}{Card background color. See \url{http://framework7.io/docs/cards.html}.}
}
\description{
Build a Framework7 expandable card
}
\note{
Does not render well.
}
\examples{
if(interactive()){
 library(shiny)
 library(shinyF7)

 shiny::shinyApp(
  ui = f7Page(
    title = "Expandable Cards",
    f7Init(theme = "auto"),
    f7ExpandableCard(
     title = "Expandable Card",
     color = "blue",
     titleColor = "white",
     subtitle = "Click on me pleaaaaase",
     "Framework7 - is a free and open source HTML mobile framework
     to develop hybrid mobile apps or web apps with iOS or Android
     native look and feel. It is also an indispensable prototyping apps tool
     to show working app prototype as soon as possible in case you need to."
    ),
    f7ExpandableCard(
     title = "Expandable Card",
     color = "green",
     titleColor = "white",
     "Framework7 - is a free and open source HTML mobile framework
     to develop hybrid mobile apps or web apps with iOS or Android
     native look and feel. It is also an indispensable prototyping apps tool
     to show working app prototype as soon as possible in case you need to."
    )
  ),
  server = function(input, output) {}
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
