% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7-fab.R
\name{f7Fab}
\alias{f7Fab}
\title{Create a Framework7 fast action button (FAB)}
\usage{
f7Fab(inputId, label, width = NULL, ...)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{label}{The contents of the button or link--usually a text label, but
you could also use any other HTML, like an image.}

\item{width}{The width of the input, e.g. \code{'400px'}, or \code{'100\%'};
see \code{\link{validateCssUnit}}.}

\item{...}{Named attributes to be applied to the button or link.}
}
\description{
Build a Framework7 fast action button (FAB)
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
