% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7Fab.R
\name{f7FabClose}
\alias{f7FabClose}
\title{Indicate that the current tag should close the \link{f7Fabs}}
\usage{
f7FabClose(tag)
}
\arguments{
\item{tag}{Target tag.}
}
\description{
Indicate that the current tag should close the \link{f7Fabs}
}
