% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7Fab.R
\name{f7FabMorphTarget}
\alias{f7FabMorphTarget}
\title{Convert a tag into a target morphing}
\usage{
f7FabMorphTarget(tag)
}
\arguments{
\item{tag}{Target tag.}
}
\description{
Convert a tag into a target morphing
}
\examples{
if (interactive()) {
 library(shiny)
 library(shinyMobile)

 shiny::shinyApp(
   ui = f7Page(
     f7SingleLayout(
       navbar = f7Navbar(title = "f7Fabs Morph"),
       toolbar = f7Toolbar(
         position = "bottom",
         lapply(1:3, function(i) f7Link(i) \%>\% f7FabClose())
       ) \%>\% f7FabMorphTarget(),
       # put an empty f7Fabs container
       f7Fabs(
         extended = TRUE,
         label = "Open",
         position = "center-top",
         color = "yellow",
         sideOpen = "right",
         morph = TRUE,
         morphTarget = ".toolbar"
       )
     )

   ),
   server = function(input, output) {}
 )
}
}
