% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7Fab.R
\name{f7Fabs}
\alias{f7Fabs}
\title{Create a Framework7 container for floating action button (FAB)}
\usage{
f7Fabs(..., position = c("right-top", "right-center", "right-bottom",
  "left-top", "left-center", "left-bottom", "center-right",
  "center-center", "center-left", "center-top", "center-bottom"),
  color = NULL, backgroundColor = "gainsboro", sideOpen = c("left",
  "right", "top", "bottom", "center"))
}
\arguments{
\item{...}{Slot for \link{f7Fab}.}

\item{position}{Container position.}

\item{color}{Container color.}

\item{backgroundColor}{Container background color. "gainsboro" by default.}

\item{sideOpen}{When the container is pressed, indicate where buttons are displayed.}
}
\description{
Build a Framework7 container for floating action button (FAB)
}
\note{
The background color might be an issue depending on the parent container. Consider
it experimental.
}
\examples{
if(interactive()){
 library(shiny)
 library(shinyF7)

 shiny::shinyApp(
  ui = f7Page(
    f7Init(theme = "auto"),
    color = "pink",
    title = "Floating action buttons",
    f7Fabs(
      position = "center-top",
      color = "yellow",
      sideOpen = "right",
      lapply(1:4, function(i) f7Fab(paste0("btn", i), i))
    ),
    lapply(1:4, function(i) verbatimTextOutput(paste0("res", i))),

    f7Fabs(
      position = "center-center",
      color = "purple",
      sideOpen = "center",
      lapply(5:8, function(i) f7Fab(paste0("btn", i), i))
    ),
    lapply(5:8, function(i) verbatimTextOutput(paste0("res", i))),

    f7Fabs(
      position = "left-bottom",
      color = "pink",
      sideOpen = "top",
      lapply(9:12, function(i) f7Fab(paste0("btn", i), i))
    )

  ),
  server = function(input, output) {
    lapply(1:12, function(i) {
      output[[paste0("res", i)]] <- renderPrint(input[[paste0("btn", i)]])
    })
  }
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
