% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7Grid.R
\name{f7Flex}
\alias{f7Flex}
\title{Create a Framework7 flex container}
\usage{
f7Flex(...)
}
\arguments{
\item{...}{Items.}
}
\description{
Build a Framework7 flex container
}
\examples{
if(interactive()){
 library(shiny)
 library(shinyF7)

 shiny::shinyApp(
   ui = f7Page(
    title = "Align",
    f7Flex(
     f7Block(strong = TRUE),
     f7Block(strong = TRUE),
     f7Block(strong = TRUE)
    )
   ),
   server = function(input, output) {}
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
