% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typography.R
\name{f7Float}
\alias{f7Float}
\title{Framework7 float utility}
\usage{
f7Float(tag, side = c("left", "right"))
}
\arguments{
\item{tag}{Tag to float.}

\item{side}{Side to float: "left" or "right".}
}
\description{
\code{f7Float} is an alignment utility for items.
}
\examples{
if(interactive()){
 library(shiny)
 library(shinyMobile)

 shinyApp(
   ui = f7Page(
    title = "Float",
    f7SingleLayout(
     navbar = f7Navbar(title = "f7Float"),
     f7Float(h1("Left"), side = "left"),
    f7Float(h1("Right"), side = "right")
    )
   ),
   server = function(input, output) {}
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
