% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7-inputs.R
\name{f7Form}
\alias{f7Form}
\title{Creates an input form}
\usage{
f7Form(..., id)
}
\arguments{
\item{...}{A list of input elements.}

\item{id}{Form unique id. Using input$\if{html}{\out{<id>}} gives
the form result.}
}
\description{
Instead of having shiny return one input
at a time, a form is a collection of related inputs.
The form returns a list with all sub-inputs as elements.
This avoids to have to deal with too many inputs.
}
\details{
This only works with elements having
an input HTML tag.
}
