% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7Searchbar.R
\name{f7Found}
\alias{f7Found}
\title{Utility to display an item when the search is successful.}
\usage{
f7Found(tag)
}
\arguments{
\item{tag}{tag to display. When using \link{f7Searchbar}, one must
wrap the items to search in inside \link{f7Found}.}
}
\description{
Use with \link{f7Searchbar}.
}
