% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7Gallery.R
\name{f7Gallery}
\alias{f7Gallery}
\title{Launch the shinyMobile Gallery}
\usage{
f7Gallery()
}
\description{
A gallery of all components available in shinyMobile.
}
\examples{
if (interactive()) {

 f7Gallery()

}
}
