% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7Grid.R
\name{f7Grid}
\alias{f7Grid}
\title{Framework7 grid container}
\usage{
f7Grid(..., cols, gap = TRUE, responsiveCl = NULL)
}
\arguments{
\item{...}{Row content.}

\item{cols}{Columns number. Numeric between 1 and 20.}

\item{gap}{Whether to display gap between columns. TRUE by default.}

\item{responsiveCl}{Class for responsive behavior.
The format must be \verb{<SIZE-NCOLS>} such as \verb{<medium-4>}, \verb{<large-3>}.
SIZE must be one of \code{c("xsmall", "small", "medium", "large", "xlarge")}.
NCOLS has to be between 1 and 20.}
}
\description{
Grid container for elements
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
