% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7-inputs.R
\name{f7GroupInput}
\alias{f7GroupInput}
\title{Framework7 group input}
\usage{
f7GroupInput(type, inputId, label, choices, selected, position, style)
}
\arguments{
\item{inputId}{Input id.}

\item{label}{Input label}

\item{choices}{List of choices. Can be a simple
vector or named list or a list of \link{f7RadioChoice} or
\link{f7CheckboxChoice}}

\item{selected}{Selected element. NULL by default. If you pass
\link{f7RadioChoice} or \link{f7CheckboxChoice} in choices,
selected must be a numeric value corresponding to the index of the element to select.}

\item{position}{Check mark side.
\code{"left"} or \code{"right"}.}

\item{style}{Input style. Inherit from \link{f7List} options
such as outline, inset, strong and dividers.}
}
\description{
Useful to create \code{f7Radio} and \link{f7CheckboxGroup}.
}
\keyword{internal}
