% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7Searchbar.R
\name{f7HideOnEnable}
\alias{f7HideOnEnable}
\title{Utility to hide a given tag when \link{f7Searchbar} is enabled.}
\usage{
f7HideOnEnable(tag)
}
\arguments{
\item{tag}{tag to hide.}
}
\description{
Use with \link{f7Searchbar}.
}
