% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7Icon.R
\name{f7Icon}
\alias{f7Icon}
\title{Create a Framework7 icon}
\usage{
f7Icon(..., lib = NULL, color = NULL)
}
\arguments{
\item{...}{Icon name and \link{f7Badge}.}

\item{lib}{Library to use: NULL, "ios" or "md". Leave NULL by default. Specify, md or ios
if you want to hide/show icons on specific devices.}

\item{color}{Icon color, if any.}
}
\description{
Build a Framework7 icon
}
\examples{
if(interactive()){
 library(shiny)
 library(shinyMobile)

 shiny::shinyApp(
  ui = f7Page(
    title = "Icons",
    init = f7Init(theme = "light", skin = "ios"),
    f7SingleLayout(
     navbar = f7Navbar(title = "icons"),
     f7List(
      f7ListItem(
        title = tagList(
         f7Icon("email")
        )
      ),
      f7ListItem(
        title = tagList(
         f7Icon("email_fill", color = "green")
        )
      ),
      f7ListItem(
        title = f7Icon("home", f7Badge("1", color = "red"))
      ),
      f7ListItem(
        title = f7Icon("email_fill", lib = "md"),
        "This will not appear since only for material design"
      )
     )
    )
  ),
  server = function(input, output) {}
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
