% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7Page.R
\name{f7Item}
\alias{f7Item}
\title{Framework7 body item}
\usage{
f7Item(..., tabName)
}
\arguments{
\item{...}{Item content.}

\item{tabName}{Item id. Must be unique, without space nor punctuation symbols.}
}
\description{
Similar to  \link{f7Tab} but for the \link{f7SplitLayout}.
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
