% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7Page.R
\name{f7Items}
\alias{f7Items}
\title{Framework7 item container}
\usage{
f7Items(...)
}
\arguments{
\item{...}{Slot for wrapper for \link{f7Item}.}
}
\description{
Build a Framework7 wrapper for \link{f7Item}
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
