% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7Card.R
\name{f7ListCard}
\alias{f7ListCard}
\title{Create a Framework7 list card}
\usage{
f7ListCard(...)
}
\arguments{
\item{...}{Slot for \link{f7ListCardItem}.}
}
\description{
Build a Framework7 list card
}
\note{
The link does not work.
}
\examples{
if(interactive()){
 library(shiny)
 library(shinyF7)

 shiny::shinyApp(
  ui = f7Page(
    title = "List Card",
    f7Init(theme = "auto"),
    f7ListCard(
     f7ListCardItem(
      url = "https://www.google.com",
      title = "Item 1"
     ),
     f7ListCardItem(
      url = "https://www.google.com",
      title = "Item 2"
     )
    )
  ),
  server = function(input, output) {}
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
