% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7Card.R
\name{f7ListCardItem}
\alias{f7ListCardItem}
\title{Create a Framework7 list card item}
\usage{
f7ListCardItem(url = NULL, title = NULL)
}
\arguments{
\item{url}{Item url.}

\item{title}{item title.}
}
\description{
Build a Framework7 list card item
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
