% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7List.R
\name{f7ListGroup}
\alias{f7ListGroup}
\title{Create a framework 7 group of contacts}
\usage{
f7ListGroup(..., title)
}
\arguments{
\item{...}{slot for \link{f7ListItem}.}

\item{title}{Group title.}
}
\description{
Create a framework 7 group of contacts
}
