% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7List.R
\name{f7ListIndexItem}
\alias{f7ListIndexItem}
\title{Create a Framework 7 list index item}
\usage{
f7ListIndexItem(...)
}
\arguments{
\item{...}{Attributes and children of the element. Named arguments become
attributes, and positional arguments become children. Valid children are
tags, single-character character vectors (which become text nodes), and raw
HTML (see \code{\link{HTML}}). You can also pass lists that contain tags,
text nodes, and HTML.}
}
\description{
Create a Framework 7 list index item
}
