% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7List.R
\name{f7ListItem}
\alias{f7ListItem}
\title{Create a Framework 7 contact item}
\usage{
f7ListItem(
  ...,
  id = NULL,
  title = NULL,
  subtitle = NULL,
  header = NULL,
  footer = NULL,
  href = NULL,
  media = NULL,
  right = NULL,
  routable = FALSE
)
}
\arguments{
\item{...}{Item text.}

\item{id}{Optional id for item.}

\item{title}{Item title.}

\item{subtitle}{Item subtitle. Only work if the \link{f7List} mode is media.}

\item{header}{Item header. Do not use when \link{f7List} mode is not NULL.}

\item{footer}{Item footer. Do not use when \link{f7List} mode is not NULL.}

\item{href}{Item external link.}

\item{media}{Expect \link{f7Icon} or \code{img}.}

\item{right}{Right content if any.}

\item{routable}{Works when href is not NULL. Default to FALSE. If TRUE,
the list item may point to another page. See \link{f7MultiLayout}.}
}
\description{
Create a Framework 7 contact item
}
