% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7Page.R
\name{f7mainPanel}
\alias{f7mainPanel}
\title{Create a Framework7 main panel container}
\usage{
f7mainPanel(..., navbar, toolbar, navbarHideScroll = FALSE)
}
\arguments{
\item{...}{Content.}

\item{navbar}{Slot for \link{f7Navbar}.}

\item{toolbar}{Slot for \link{f7Toolbar}.}

\item{navbarHideScroll}{Whether to hide the navbar on scroll. FALSE by default.}
}
\description{
Build a Framework7 main panel container
}
\examples{
if(interactive()){
 library(shiny)
 library(shinyF7)

}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
