% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7Card.R
\name{f7MediaCard}
\alias{f7MediaCard}
\title{Create a Framework7 media card}
\usage{
f7MediaCard(..., title = NULL, footer = NULL)
}
\arguments{
\item{...}{Slot for \link{f7MediaCardItem}.}

\item{title}{Card title.}

\item{footer}{Footer content, if any. Must be wrapped in a tagList.}
}
\description{
Build a Framework7 media card
}
\examples{
if(interactive()){
 library(shiny)
 library(shinyF7)

 shiny::shinyApp(
  ui = f7Page(
    title = "Media Card",
    f7Init(theme = "auto"),
    f7SingleLayout(
     navbar = f7Navbar(title = "f7MediaCard"),
     f7MediaCard(
     title = "New Releases:",
     f7MediaCardItem(
      src = "http://lorempixel.com/88/88/fashion/4",
      title = "Yellow Submarine",
      subtitle = "Beatles"
     ),
     f7MediaCardItem(
      src = "http://lorempixel.com/88/88/fashion/5",
      title = "Don't Stop Me Now",
      subtitle = "Queen"
     ),
     footer = tagList(
      span("January 20", 2015),
      span(5, "comments")
     )
    )
    )
  ),
  server = function(input, output) {}
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
