% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7Card.R
\name{f7MediaCardItem}
\alias{f7MediaCardItem}
\title{Create a Framework7 media card item}
\usage{
f7MediaCardItem(src = NULL, title = NULL, subtitle = NULL)
}
\arguments{
\item{src}{Item image url or path.}

\item{title}{Item title.}

\item{subtitle}{Item subtitle.}
}
\description{
Build a Framework7 media card item
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
