% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7Menu.R
\name{f7Menu}
\alias{f7Menu}
\title{Create a framework7 menu}
\usage{
f7Menu(..., id = NULL)
}
\arguments{
\item{...}{Slot for \link{f7MenuItem} or \link{f7MenuDropown}.}

\item{id}{Menu id. This is required when once wants to programmatically toggle
the menu on the server side with \link{f7ToggleMenu}.}
}
\description{
Create a framework7 menu
}
\examples{
if (interactive()) {
 library(shiny)
 library(shinyMobile)


}
}
