% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7Menu.R
\name{f7MenuDropdown}
\alias{f7MenuDropdown}
\title{Create dropdown menu for \link{f7Menu}.}
\usage{
f7MenuDropdown(..., label, side = c("left", "center", "right"))
}
\arguments{
\item{...}{Slot for \link{f7MenuItem} and \link{f7MenuDropdownDivider}.}

\item{label}{Button label.}

\item{side}{Dropdown opening side. Choose among \code{c("left", "center", "right")}.}
}
\description{
Create dropdown menu for \link{f7Menu}.
}
