% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7Menu.R
\name{f7MenuItem}
\alias{f7MenuItem}
\title{Create a special action button for \link{f7Menu}.}
\usage{
f7MenuItem(inputId, label)
}
\arguments{
\item{inputId}{Input id.}

\item{label}{Button label.}
}
\description{
Create a special action button for \link{f7Menu}.
}
