% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7Messages.R
\name{f7Message}
\alias{f7Message}
\title{Create a f7Message}
\usage{
f7Message(
  text,
  name,
  type = c("sent", "received"),
  header = NULL,
  footer = NULL,
  avatar = NULL,
  textHeader = NULL,
  textFooter = NULL,
  image = NULL,
  imageSrc = NULL,
  cssClass = NULL
)
}
\arguments{
\item{text}{Message text.}

\item{name}{Sender name.}

\item{type}{Message type - sent or received.}

\item{header}{Single message header.}

\item{footer}{Single message footer.}

\item{avatar}{Sender avatar URL string.}

\item{textHeader}{Message text header.}

\item{textFooter}{Message text footer.}

\item{image}{Message image HTML string, e.g. <img src="path/to/image">. Can be used instead of imageSrc parameter.}

\item{imageSrc}{Message image URL string. Can be used instead of image parameter.}

\item{cssClass}{Additional CSS class to set on message HTML element.}
}
\description{
Create a f7Message
}
