% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7Messages.R
\name{f7MessageBar}
\alias{f7MessageBar}
\title{Create a f7MessageBar to add new messages}
\usage{
f7MessageBar(inputId, placeholder = "Message")
}
\arguments{
\item{inputId}{Unique id.}

\item{placeholder}{Textarea placeholder.}
}
\description{
Insert before \link{f7Messages}. See examples.
}
