% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7Messages.R
\name{f7Messages}
\alias{f7Messages}
\title{Create a Framework7 messages container}
\usage{
f7Messages(
  id,
  title = NULL,
  autoLayout = TRUE,
  newMessagesFirst = FALSE,
  scrollMessages = TRUE,
  scrollMessagesOnEdge = TRUE
)
}
\arguments{
\item{id}{Container id.}

\item{title}{Optional messages title.}

\item{autoLayout}{Enable Auto Layout to add all required additional classes
automatically based on passed conditions.}

\item{newMessagesFirst}{Enable if you want to use new messages on top,
instead of having them on bottom.}

\item{scrollMessages}{Enable/disable messages autoscrolling when adding new message.}

\item{scrollMessagesOnEdge}{If enabled then messages autoscrolling will happen only
when user is on top/bottom of the messages view.}
}
\description{
Create a Framework7 messages container
}
\examples{
if (interactive()) {
 library(shiny)
 library(shinyMobile)
 shinyApp(
   ui = f7Page(
     preloader = FALSE,
     title = "My app",
     f7SingleLayout(
       navbar = f7Navbar(title = "f7Messages"),
       f7Messages(id = "messagelist")
     )
   ),
   server = function(input, output, session) {
   }
 )
}
}
