% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7Navbar.R
\name{f7Navbar}
\alias{f7Navbar}
\title{Create a Framework7 Navbar}
\usage{
f7Navbar(hairline = TRUE, shadow = TRUE)
}
\arguments{
\item{hairline}{Whether to display a thin border on the top of the navbar. TRUE by default.}

\item{shadow}{Whether to display a shadow. TRUE by default.}
}
\description{
Build a Framework7 Navbar
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
