% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7Appbar.R
\name{f7Next}
\alias{f7Next}
\title{Create a framework 7 next button}
\usage{
f7Next(targetId)
}
\arguments{
\item{targetId}{\link{f7Tabs} id.}
}
\description{
This buttons allows to switch between multiple \link{f7Tab}.
}
