% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7Searchbar.R
\name{f7NotFound}
\alias{f7NotFound}
\title{Utility to display an item when the search is unsuccessful.}
\usage{
f7NotFound(tag)
}
\arguments{
\item{tag}{tag to use.}
}
\description{
Use with \link{f7Searchbar}.
}
