% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7Notif.R
\name{f7Notif}
\alias{f7Notif}
\title{Create a Framework7 notification}
\usage{
f7Notif(text, title, titleRightText, subtitle = NULL,
  closeTimeout = 5000, closeButton = TRUE, session)
}
\arguments{
\item{text}{Notification content.}

\item{title}{Notification title.}

\item{titleRightText}{Notification right text.}

\item{subtitle}{Notification subtitle}

\item{closeTimeout}{Time before notification close.}

\item{closeButton}{Whether to display a close button.
FALSE by default.}

\item{session}{shiny session.}
}
\description{
Create a Framework7 notification
}
\examples{
if (interactive()) {
  library(shiny)
  library(shinyF7)
  shinyApp(
    ui = f7Page(
      color = "pink",
      title = "My app",
      init = f7Init(theme = "ios"),
      f7SingleLayout(
        navbar = f7Navbar(title = "f7Notif"),
        f7Button(inputId = "goButton", "Go!")
      )
    ),
    server = function(input, output, session) {
      shiny::observeEvent(input$goButton,{
        f7Notif(
          text = "test",
          title = "Notification",
          subtitle = "A subtitle",
          titleRightText = "now",
          session = session
        )
      })
    }
  )
}
}
