% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7Page.R
\name{f7Page}
\alias{f7Page}
\title{Create a Framework7 page}
\usage{
f7Page(..., title = NULL, navbar = NULL, navbarHideScroll = FALSE,
  dark_mode = FALSE)
}
\arguments{
\item{...}{Any element. They are inserted in a grid. Use the shiny fluidRow function
to create a row and insert metroUiCol inside. The maximum with is 12
(3 columns or lenght 4, 4 columns of lenght 3, ...).}

\item{title}{Page title.}

\item{navbar}{Slot for \link{f7Navbar}.}

\item{navbarHideScroll}{Whether to hide the navbar on scroll. FALSE by default.}

\item{dark_mode}{Whether to enable the dark mode. FALSE by default.}
}
\description{
Build a Framework7 page
}
\examples{
if(interactive()){
 library(shiny)
 library(shinyF7)

 shiny::shinyApp(
   ui = f7Page(title = "My app"),
   server = function(input, output) {}
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
