% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7Panel.R
\name{f7Panel}
\alias{f7Panel}
\title{Create a Framework7 panel}
\usage{
f7Panel(..., title = NULL, side = c("left", "right"),
  theme = c("dark", "light"), style = c("reveal", "cover"))
}
\arguments{
\item{...}{Panel content. Also include \link{f7PanelMenu} and \link{f7PanelItem}.}

\item{title}{Panel title.}

\item{side}{Panel side: "left" or "right".}

\item{theme}{Panel background color: "dark" or "light".}

\item{style}{Whether the panel should behave when opened: "cover" or "reveal".}
}
\description{
Build a Framework7 panel
}
\note{
When style is "reveal", makes the app freeze.
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
