% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7Panel.R
\name{f7PanelItem}
\alias{f7PanelItem}
\title{Create a Framework7 sidebar menu item}
\usage{
f7PanelItem(title, tabName, icon = NULL, active = FALSE)
}
\arguments{
\item{title}{Item name.}

\item{tabName}{Item unique tabName. Must correspond to what is passed to
\link{f7Item}.}

\item{icon}{Item icon.}

\item{active}{Whether the item is active at start. Default to FALSE.}
}
\description{
Build a Framework7 sidebar menu item for \link{f7SplitLayout}.
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
