% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7Panel.R
\name{f7PanelMenu}
\alias{f7PanelMenu}
\title{Create a Framework7 sidebar menu}
\usage{
f7PanelMenu(..., id = NULL)
}
\arguments{
\item{...}{Slot for \link{f7PanelItem}.}

\item{id}{Unique id to access the currently selected item.}
}
\description{
Build a Framework7 sidebar menu
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
