% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7-inputs.R
\name{f7pickerInput}
\alias{f7pickerInput}
\title{Create a Framework7 picker input}
\usage{
f7pickerInput(inputId, label, placeHolder = NULL, choices)
}
\arguments{
\item{inputId}{Picker input id.}

\item{label}{Picker label.}

\item{placeHolder}{Text to write in the container.}

\item{choices}{Picker choices.}
}
\description{
Build a Framework7 picker input
}
\examples{
if(interactive()){
 library(shiny)
 library(shinyF7)

 shiny::shinyApp(
   ui = f7Page(
    title = "My app",
    f7pickerInput(
     inputId = "my-picker",
     placeHolder = "Some text here!",
     label = "Picker Input",
     choices = c('a', 'b', 'c')
    )
   ),
   server = function(input, output) {}
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
