% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7Popover.R
\name{f7PopoverTarget}
\alias{f7PopoverTarget}
\title{Define a popover target}
\usage{
f7PopoverTarget(tag, targetId)
}
\arguments{
\item{tag}{Tag that will be targeted. Must be a f7Input element.}

\item{targetId}{Popover id. Must correspond to the \link{f7Popover}targetId.}
}
\description{
This must be used in combination of \link{f7Popover}.
Only works for input elements!
}
