% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7Popup.R
\name{f7Popup}
\alias{f7Popup}
\title{Create a f7 popup}
\usage{
f7Popup(..., id, label = "Open", title)
}
\arguments{
\item{...}{Content.}

\item{id}{Popup unique id.}

\item{label}{Popup trigger label.}

\item{title}{Title.}
}
\description{
Create a f7 popup
}
\examples{
if (interactive()) {
 library(shiny)
 library(shinyF7)
 shiny::shinyApp(
   ui = f7Page(
     color = "pink",
     title = "My app",
     f7Init("ios"),
     f7SingleLayout(
      navbar = f7Navbar(
        title = "Single Layout",
        hairline = FALSE,
        shadow = TRUE
      ),
      f7Popup(
       id = "popup1",
       label = "Open",
       title = "My first popup",
       "Lorem ipsum dolor sit amet, consectetur adipiscing elit.
         Quisque ac diam ac quam euismod porta vel a nunc. Quisque sodales
         scelerisque est, at porta justo cursus ac"
      )
     )
   ),
   server = function(input, output) {}
 )
}
}
