% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7Progress.R
\name{f7Progress}
\alias{f7Progress}
\title{Create a Framework7 progress bar}
\usage{
f7Progress(value, status = NULL)
}
\arguments{
\item{value}{Progress value. Between 0 and 100.}

\item{status}{Progress color. See \url{http://framework7.io/docs/progressbar.html}.}
}
\description{
Build a Framework7 progress bar
}
\note{
Buggy display
}
\examples{
if(interactive()){
 library(shiny)
 library(shinyF7)

 shiny::shinyApp(
   ui = f7Page(
    title = "My app",
    f7Card(
     f7Progress(value = 10),
     f7Progress(value = 10, status = "red")
    )
   ),
   server = function(input, output) {}
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
