% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7Progress.R
\name{f7ProgressInf}
\alias{f7ProgressInf}
\title{Create a Framework7 infinite progress bar}
\usage{
f7ProgressInf(status = NULL)
}
\arguments{
\item{status}{Progress color. See \url{http://framework7.io/docs/progressbar.html}.}
}
\description{
Build a Framework7 infinite progress bar
}
\note{
Buggy display
}
\examples{
if(interactive()){
 library(shiny)
 library(shinyF7)

 shiny::shinyApp(
   ui = f7Page(
    title = "My app",
    f7Card(
     f7ProgressInf()
    )
   ),
   server = function(input, output) {}
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
