% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7PullToRefresh.R
\name{f7PullToRefresh}
\alias{f7PullToRefresh}
\title{Create a Framework7 pull to refresh event}
\usage{
f7PullToRefresh(id, bottom = FALSE, closeTimeout = 2000, session)
}
\arguments{
\item{id}{Unique ptr id.}

\item{bottom}{Whether to pull from bottom. FALSE by default.}

\item{closeTimeout}{Time before the ptr is done.}

\item{session}{Shiny session object}
}
\description{
Triggered after scroll down or up
}
\examples{
if (interactive()) {
 library(shiny)
 library(shinyF7)

}
}
