% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7-deprecated.R
\name{f7Row}
\alias{f7Row}
\title{Framework7 row container}
\usage{
f7Row(..., gap = TRUE)
}
\arguments{
\item{...}{Row content.}

\item{gap}{Whether to display gap between columns. TRUE by default.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\link{f7Grid} is a replacement
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
\keyword{internal}
