% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7Searchbar.R
\name{f7SearchIgnore}
\alias{f7SearchIgnore}
\title{Utility to ignore an item from search.}
\usage{
f7SearchIgnore(tag)
}
\arguments{
\item{tag}{tag to ignore.}
}
\description{
Use with \link{f7Searchbar}.
}
