% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7Button.R
\name{f7Segment}
\alias{f7Segment}
\title{Create a Framework7 segmented button container}
\usage{
f7Segment(..., container = c("segment", "row"), shadow = FALSE,
  rounded = FALSE)
}
\arguments{
\item{...}{Slot for \link{f7Button}.}

\item{container}{Either "row" or "segment".}

\item{shadow}{Button shadow. FALSE by default. Only for material design and if the container is segment.}

\item{rounded}{Round style. FALSE by default. Only if the container is segment.}
}
\description{
Build a Framework7 segmented button container
}
\examples{
if(interactive()){
 library(shiny)
 library(shinyF7)

 shiny::shinyApp(
  ui = f7Page(
    title = "Button Segments",
    f7Init(theme = "md"),
    f7Segment(
     rounded = TRUE,
     container = "segment",
     f7Button(color = "red", label = "Action Button", inputId = "button2", fill = TRUE),
     f7Button(color = "green", label = "My button", outline = TRUE),
     f7Button(color = "yellow", label = "My button", rounded = TRUE)
    ),
    f7Segment(
     container = "row",
     f7Button(color = "pink", label = "My button", shadow = TRUE),
     f7Button(color = "purple", label = "My button", size = "large"),
     f7Button(color = "orange", label = "My button", size = "small")
    ),

    verbatimTextOutput("val")
  ),
  server = function(input, output) {
   output$val <- renderPrint(input$button2)
  }
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
