% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7-deprecated.R
\name{f7Shadow}
\alias{f7Shadow}
\title{Framework7 shadow effect}
\usage{
f7Shadow(tag, intensity, hover = FALSE, pressed = FALSE)
}
\arguments{
\item{tag}{Tag to apply the shadow on.}

\item{intensity}{Shadow intensity. Numeric between 1 and 24. 24 is the highest elevation.}

\item{hover}{Whether to display the shadow on hover. FALSE by default.}

\item{pressed}{Whether to display the shadow on click. FALSE by default.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}.
Creates a shadow effect to apply on UI elements like \link{f7Card}.
}
\keyword{internal}
