% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7Sheet.R
\name{f7Sheet}
\alias{f7Sheet}
\title{Create an f7 sheet modal}
\usage{
f7Sheet(..., id, label = "Open", orientation = c("top", "bottom"))
}
\arguments{
\item{...}{Sheet content.}

\item{id}{Sheet unique id.}

\item{label}{Trigger label.}

\item{orientation}{"top" or "bottom".}
}
\description{
Create an f7 sheet modal
}
\examples{
if (interactive()) {
 library(shiny)
 library(shinyF7)
  shiny::shinyApp(
    ui = f7Page(
      color = "pink",
      title = "My app",
      f7Init("ios"),
      f7Sheet(
        id = "sheet1",
        label = "More",
        orientation = "bottom",
        "Lorem ipsum dolor sit amet, consectetur adipiscing elit.
        Quisque ac diam ac quam euismod porta vel a nunc. Quisque sodales
        scelerisque est, at porta justo cursus ac"
      )
    ),
    server = function(input, output) {}
  )
}
}
