% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7Page.R
\name{f7Sidebar}
\alias{f7Sidebar}
\title{Create a Framework7 sidebar}
\usage{
f7Sidebar(..., title = NULL, side = c("left", "right"),
  theme = c("dark", "light"))
}
\arguments{
\item{...}{Panel content.}

\item{title}{Panel title.}

\item{side}{Panel side: "left" or "right".}

\item{theme}{Panel background color: "dark" or "light".}
}
\description{
Build a Framework7 sidebar
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
