% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7Page.R
\name{f7SidebarLayout}
\alias{f7SidebarLayout}
\title{Create a Framework7 page with split layout}
\usage{
f7SidebarLayout(sidebarPanel, mainPanel)
}
\arguments{
\item{sidebarPanel}{Slot for \link{f7Sidebar}.}

\item{mainPanel}{Slot for \link{f7mainPanel}.}
}
\description{
Build a Framework7 page with split layout
}
\examples{
if(interactive()){
 library(shiny)
 library(shinyF7)

}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
