% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7Swiper.R
\name{f7Slide}
\alias{f7Slide}
\title{Create a Framework7 slide}
\usage{
f7Slide(...)
}
\arguments{
\item{...}{Any element.}
}
\description{
Build a Framework7 slide
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
