% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7-deprecated.R
\name{f7SocialCard}
\alias{f7SocialCard}
\title{Framework7 social card}
\usage{
f7SocialCard(..., image = NULL, author = NULL, date = NULL, footer = NULL)
}
\arguments{
\item{...}{Card content.}

\item{image}{Author img.}

\item{author}{Author.}

\item{date}{Date.}

\item{footer}{Footer content, if any. Must be wrapped in a tagList.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}.
\code{f7SocialCard} is a special card for social content.
}
\keyword{internal}
