% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7Swipeout.R
\name{f7SwipeoutItem}
\alias{f7SwipeoutItem}
\title{Create a framework7 swipeout item}
\usage{
f7SwipeoutItem(id, label, color = NULL)
}
\arguments{
\item{id}{Item unique id.}

\item{label}{Item label.}

\item{color}{Item color.}
}
\description{
Insert in \link{f7Swipeout}
}
