% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7Swiper.R
\name{f7Swiper}
\alias{f7Swiper}
\title{Create a Framework7 swiper}
\usage{
f7Swiper(..., pagination = TRUE, space = 50, n_slides = "auto",
  centered = TRUE, vertical = FALSE, speed = 900)
}
\arguments{
\item{...}{Slot for \link{f7Slide}.}

\item{pagination}{Whether to show pagination. TRUE by default.}

\item{space}{Space between slides. 50 by default. Only if pagination is TRUE.}

\item{n_slides}{Number of slides at a time. Only if pagination is TRUE. Set to "auto" by default.}

\item{centered}{Whether to center slides. Only if pagination is TRUE.}

\item{vertical}{Whether to display slides vertically. FALSE by default. Only if pagination is TRUE.}

\item{speed}{Slides speed. Numeric.}
}
\description{
Build a Framework7 swiper (like carousel)
}
\note{
Does not work
}
\examples{
if(interactive()){
 library(shiny)
 library(shinyF7)

 shiny::shinyApp(
   ui = f7Page(
    title = "My app",
    f7Init(theme = "auto"),
    actionButton("newplot", "New plot"),
    f7Swiper(
     f7Slide("Slide 1"),
     f7Slide(plotOutput("plot"))
    )
   ),
   server = function(input, output) {
     output$plot <- renderPlot({
      input$newplot
      # Add a little noise to the cars data
      cars2 <- cars + rnorm(nrow(cars))
      plot(cars2)
     })
   }
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
