% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7-tabs.R
\name{f7Tab}
\alias{f7Tab}
\title{Create a Framework7 tab item}
\usage{
f7Tab(..., tabName, icon = NULL, active = FALSE, hidden = FALSE)
}
\arguments{
\item{...}{Item content.}

\item{tabName}{Item id. Must be unique.}

\item{icon}{Item icon. Expect \link{f7Icon} function with the suitable lib argument
(either md or ios or NULL for native f7 icons).}

\item{active}{Whether the tab is active at start. Do not select multiple tabs, only
the first one will be set to active.}

\item{hidden}{Whether to hide the tab. This is useful when you want to add invisible tabs
(that do not appear in the tabbar) but you can still navigate with \link{updateF7Tabs}.}
}
\description{
Build a Framework7 tab item
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
