% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7Page.R
\name{f7TabLayout}
\alias{f7TabLayout}
\title{Create a Framework7 page with tab layout}
\usage{
f7TabLayout(...)
}
\arguments{
\item{...}{Slot for \link{f7Tabs} and/or \link{f7Navbar} and/or \link{f7Panel}.}
}
\description{
Build a Framework7 page with tab layout
}
\examples{
if(interactive()){
 library(shiny)
 library(shinyF7)

}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
