% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7-tabs.R
\name{f7TabLink}
\alias{f7TabLink}
\title{Special button/link to insert in the tabbar}
\usage{
f7TabLink(..., icon = NULL, label = NULL)
}
\arguments{
\item{...}{Any attribute like \code{`data-sheet`}, id, ...}

\item{icon}{Expect \link{f7Icon}.}

\item{label}{Button label.}
}
\description{
Use in the .items slot of \link{f7Tabs}.
}
