% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7-tabs.R
\name{f7TabList}
\alias{f7TabList}
\title{Create a Framework7 tab list}
\usage{
f7TabList(..., bottom = FALSE, hairline = TRUE, shadow = TRUE,
  tabs = TRUE, labels = FALSE, scrollable = FALSE,
  mode = c("animated", "swipeable"))
}
\arguments{
\item{...}{Slot for f7ToolbarItem.}

\item{bottom}{Toolbar position. FALSE by default, the toolbar is displayed
below the navbar. If TRUE, it is displayed at the bottom of the page.}

\item{hairline}{Whether to display a thin border on the top of the toolbar. TRUE by default.}

\item{shadow}{Whether to display a shadow. TRUE by default.}

\item{tabs}{Whether to use the toolbar with tabs. See f7TabList and f7Tab. FALSE
by default.}

\item{labels}{Whether to use icons instead of text. Only use when tabbar is TRUE.}

\item{scrollable}{Whether to allow scrolling. FALSE by default.}

\item{mode}{Tab animation mode. Either "animated" or "swipeable".}
}
\description{
Build a Framework7 tab list
}
\examples{
if(interactive()){
 library(shiny)
 library(shinyF7)

 shiny::shinyApp(
   ui = f7Page(
    title = "My app",
    f7TabList(
     tabs = TRUE,
     mode = "swipeable",
     f7Tab(
      active = TRUE,
      id = "tab1",
      "Tab 1"
     ),
     f7Tab(
      id = "tab2",
      "Tab 2"
     ),
     f7Tab(
      id = "tab3",
      "Tab 3"
     )
    )
   ),
   server = function(input, output) {}
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
