% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7Menu.R
\name{f7ToggleMenu}
\alias{f7ToggleMenu}
\title{Toggle \link{f7Menu} on the server side}
\usage{
f7ToggleMenu(id, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{id}{Menu to target.}

\item{session}{Shiny session object.}
}
\description{
Toggle \link{f7Menu} on the server side
}
