% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7Popup.R
\name{f7TogglePopup}
\alias{f7TogglePopup}
\title{Toggle \link{f7Popup}.}
\usage{
f7TogglePopup(id, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{id}{Popup id.}

\item{session}{Shiny session.}
}
\description{
Toggle \link{f7Popup}.
}
