% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7-toolbars.R
\name{f7ToolbarItem}
\alias{f7ToolbarItem}
\title{Create a Framework7 toolbar/tabbar item}
\usage{
f7ToolbarItem(name, href)
}
\arguments{
\item{name}{Item name.}

\item{href}{Item link to. Should be related to the unique id of the tab if
it is used in a tabbar.}
}
\description{
Build a Framework7 toolbar/tabbar item
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
