% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7Treeview.R
\name{f7TreeviewGroup}
\alias{f7TreeviewGroup}
\title{Create a Framework 7 group Treeview items}
\usage{
f7TreeviewGroup(..., title, icon)
}
\arguments{
\item{...}{slot for \link{f7TreeviewItem}.}

\item{title}{Group title.}

\item{icon}{Expect \link{f7Icon}.}
}
\description{
Create a Framework 7 group Treeview items
}
