% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7Treeview.R
\name{f7TreeviewItem}
\alias{f7TreeviewItem}
\title{Create a Framework 7 Treeview item}
\usage{
f7TreeviewItem(label, icon = NULL, href = NULL)
}
\arguments{
\item{label}{Item label}

\item{icon}{Expect \link{f7Icon}.}

\item{href}{Item external link.}
}
\description{
Create a Framework 7 Treeview item
}
