% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7List.R
\name{f7VirtualList}
\alias{f7VirtualList}
\title{High performance list component}
\usage{
f7VirtualList(
  id,
  items,
  rowsBefore = NULL,
  rowsAfter = NULL,
  itemTemplate,
  cache = TRUE
)
}
\arguments{
\item{id}{Virtual list unique id.}

\item{items}{List items.}

\item{rowsBefore}{Amount of rows (items) to be rendered before current
screen scroll position. By default it is equal to double amount of
rows (items) that fit to screen.}

\item{rowsAfter}{Amount of rows (items) to be rendered after current
screen scroll position. By default it is equal to the amount of rows
(items) that fit to screen.}

\item{itemTemplate}{Template7 string template or Template7 compiled
template that used to render single item. Template should contain full HTML
layout for single item, including wrapping <li></li> tags}

\item{cache}{Disable or enable DOM cache for already rendered list items.
In this case each item will be rendered only once and all futher
manipulations will be with DOM element. It is useful if your list
items have some user interaction elements (like form elements or swipe outs)
or could be modified.}
}
\description{
Use if many components in \link{f7List}
}
\examples{
if (interactive()) {
 library(shiny)
 library(shinyMobile)

}
}
