% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7Fab.R
\name{f7Fab}
\alias{f7Fab}
\alias{updateF7Fab}
\title{Framework7 floating action button (FAB)}
\usage{
f7Fab(inputId, label, width = NULL, ..., flag = NULL)

updateF7Fab(inputId, label = NULL, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{label}{The contents of the button or link--usually a text label, but
you could also use any other HTML, like an image.}

\item{width}{The width of the input, e.g. \code{'400px'}, or \code{'100\%'};
see \code{\link[shiny:validateCssUnit]{validateCssUnit()}}.}

\item{...}{Named attributes to be applied to the button or link.}

\item{flag}{Additional text displayed next to the button content. Only works
if \link{f7Fabs} position parameter is not starting with center-...}

\item{session}{The Shiny session object, usually the default value will suffice.}
}
\description{
\code{f7Fab} generates a nice button to be put in \link{f7Fabs}.

\code{updateF7Fab} changes the label of an \link{f7Fab} input on the client.
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
