% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getF7Colors}
\alias{getF7Colors}
\title{Function to get all colors available in shinyMobile}
\usage{
getF7Colors()
}
\value{
A vector containing colors
}
\description{
Function to get all colors available in shinyMobile
}
