% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7Preloader.R
\name{hideF7Preloader}
\alias{hideF7Preloader}
\title{Framework7 preloader}
\usage{
hideF7Preloader(target = NULL, session = shiny::getDefaultReactiveDomain())
}
\description{
\code{hideF7Preloader} hides a preloader.
Use \link{f7HidePreloader} instead
}
\keyword{internal}
