% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7-deprecated.R
\name{f7Menu}
\alias{f7Menu}
\alias{f7MenuItem}
\alias{f7MenuDropdown}
\alias{f7MenuDropdownDivider}
\alias{updateF7MenuDropdown}
\title{Framework7 menu container}
\usage{
f7Menu(...)

f7MenuItem(inputId, label)

f7MenuDropdown(..., id = NULL, label, side = c("left", "center", "right"))

f7MenuDropdownDivider()

updateF7MenuDropdown(id, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{...}{Slot for \link{f7MenuItem} and \link{f7MenuDropdownDivider}.}

\item{inputId}{Menu item input id.}

\item{label}{Button label.}

\item{id}{Menu to target.}

\item{side}{Dropdown opening side. Choose among \code{c("left", "center", "right")}.}

\item{session}{Shiny session object.}
}
\description{
\code{f7Menu} is a container for \link{f7MenuItem} and/or \link{f7MenuDropdown}.

\code{f7MenuItem} creates a special action button for \link{f7Menu}.

\code{f7MenuDropdown} creates a dropdown menu for \link{f7Menu}.

\code{f7MenuDropdownDivider} creates a dropdown divider for \link{f7MenuDropdown}.

\code{updateF7MenuDropdown} toggles \link{f7MenuDropdown} on the client.
}
\keyword{internal}
