% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7-tabs.R
\name{removeF7Tab}
\alias{removeF7Tab}
\title{Framework7 tab deletion}
\usage{
removeF7Tab(id, target, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{id}{\link{f7Tabs} id.}

\item{target}{\link{f7Tab} to remove.}

\item{session}{Shiny session object.}
}
\description{
\code{removeF7Tab} removes an \link{f7Tab} in a \link{f7Tabs}.
}
\seealso{
\link{f7Tabs}
}
