% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setTheme.R
\name{setTheme}
\alias{setTheme}
\title{Custom theme setup}
\usage{
setTheme(theme = c("ios", "md", "auto"))
}
\arguments{
\item{theme}{App theme: "ios", "md", "auto".}
}
\description{
Custom theme setup
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
