% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyMobile-package.R
\docType{package}
\name{shinyMobile-package}
\alias{shinyMobile}
\alias{shinyMobile-package}
\title{shinyMobile: Mobile Ready 'shiny' Apps with Standalone Capabilities}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Develop outstanding 'shiny' apps for 'iOS' and 'Android' as well as beautiful 'shiny' gadgets. 'shinyMobile' is built on top of the latest 'Framework7' template \url{https://framework7.io}. Discover 14 new input widgets (sliders, vertical sliders, stepper, grouped action buttons, toggles, picker, smart select, ...), 2 themes (light and dark), 12 new widgets (expandable cards, badges, chips, timelines, gauges, progress bars, ...) combined with the power of server-side notifications such as alerts, modals, toasts, action sheets, sheets (and more) as well as 3 layouts (single, tabs and split).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/RinteRface/shinyMobile}
  \item \url{https://rinterface.github.io/shinyMobile/}
  \item Report bugs at \url{https://github.com/RinteRface/shinyMobile/issues}
}

}
\author{
\strong{Maintainer}: David Granjon \email{dgranjon@ymail.com}

Authors:
\itemize{
  \item Victor Perrier \email{info@dreamrs.fr}
  \item Isabelle Rudolf
}

Other contributors:
\itemize{
  \item John Coene \email{jcoenep@gmail.com} [contributor]
  \item Dieter Menne \email{dieter.menne@menne-biomed.de} [contributor]
  \item  Marvelapp (device.css wrappers) [contributor, copyright holder]
  \item Vladimir Kharlampidi \email{nolimits4web@gmail.com} (Framework7 HTML template) [contributor, copyright holder]
}

}
\keyword{internal}
