% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7-update-inputs.R
\name{updateF7Button}
\alias{updateF7Button}
\title{Change the value of a button input on the client}
\usage{
updateF7Button(session, inputId, label = NULL)
}
\arguments{
\item{session}{The session object passed to function given to the server.}

\item{inputId}{The id of the input object.}

\item{label}{The label to set for the input object.}
}
\description{
Change the value of a button input on the client
}
\examples{
if (interactive()) {
 library(shiny)
 library(shinyMobile)

 shiny::shinyApp(
  ui = f7Page(
    title = "Update f7Button",
    f7SingleLayout(
      navbar = f7Navbar(title = "Icons"),
      f7Button("btn", "Button"),
      f7Toggle("updateBtn", "Update Button")
    )
  ),
  server = function(input, output, session) {
    observe(print(input$btn))
    observeEvent(input$updateBtn, {
      if (input$updateBtn) {
        updateF7Button(
         session = session,
         inputId = "btn",
         label = "Updated"
        )
      }
    })
  }
 )
}
}
