% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7-update-inputs.R
\name{updateF7Date}
\alias{updateF7Date}
\title{Change the value of a date input on the client}
\usage{
updateF7Date(session, inputId, label = NULL, value = NULL,
  placeholder = NULL)
}
\arguments{
\item{session}{The session object passed to function given to the server.}

\item{inputId}{The id of the input object.}

\item{label}{The label to set for the input object.}

\item{value}{The value to set for the input object.}

\item{placeholder}{The placeholder to set for the input object.}
}
\description{
Change the value of a date input on the client
}
\examples{
if (interactive()) {
 library(shiny)
 library(shinyMobile)

 ui <- f7Page(
   f7SingleLayout(
    navbar = f7Navbar(title = "updateF7Text"),
    f7Fab("trigger", "Click me"),
    f7Date(inputId = "date", label = "Date", value = "2014-04-30"),
    verbatimTextOutput("datevalue")
   )
 )

 server <- function(input, output, session) {
   output$datevalue <- renderPrint(input$date)
   observeEvent(input$trigger, {
     updateF7Date(session, "date", value = "2020-01-15")
   })
 }
 shinyApp(ui, server)
}
}
