% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7-update-inputs.R
\name{updateF7Picker}
\alias{updateF7Picker}
\title{Change the value of a stepper input on the client}
\usage{
updateF7Picker(session, inputId, placeholder = NULL, value = NULL,
  choices = NULL)
}
\arguments{
\item{session}{The session object passed to function given to the server.}

\item{inputId}{The id of the input object.}

\item{placeholder}{New picker placeholder.}

\item{value}{Picker initial value, if any.}

\item{choices}{New picker choices.}
}
\description{
Change the value of a stepper input on the client
}
\examples{
if (interactive()) {
 library(shiny)
 library(shinyF7)

}
}
