% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7Page.R
\name{updateF7Routes}
\alias{updateF7Routes}
\title{Update routes on the server}
\usage{
updateF7Routes(routes, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{routes}{New list of routes.}

\item{session}{Shiny session object.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Add a route to existing app routes.
}
