% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7List.R
\name{updateF7VirtualList}
\alias{updateF7VirtualList}
\title{Update a \link{f7VirtualList} on the server side}
\usage{
updateF7VirtualList(
  id,
  action = c("appendItem", "appendItems", "prependItem", "prependItems", "replaceItem",
    "replaceAllItems", "moveItem", "insertItemBefore", "filterItems", "deleteItem",
    "deleteItems", "deleteAllItems", "scrollToItem", "clearCache"),
  item = NULL,
  items = NULL,
  index = NULL,
  indexes = NULL,
  old_index = NULL,
  new_index = NULL,
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{id}{\link{f7VirtualList} to update.}

\item{action}{Action to perform. See \url{https://framework7.io/docs/virtual-list.html}.}

\item{item}{If action is one of appendItem, prependItem, replaceItem, insertItemBefore.}

\item{items}{If action is one of appendItems, prependItems, replaceAllItems.}

\item{index}{If action is one of replaceItem, insertItemBefore, deleteItem.}

\item{indexes}{If action if one of filterItems, deleteItems.}

\item{old_index}{If action is moveItem.}

\item{new_index}{If action is moveItem.}

\item{session}{Shiny session.}
}
\description{
This function wraps all methods from \url{https://framework7.io/docs/virtual-list.html}
}
\examples{
if (interactive()) {
 library(shiny)
 library(shinyMobile)

}
}
