% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7-tabs.R
\name{validate_tabName}
\alias{validate_tabName}
\title{Validate a tab name}
\usage{
validate_tabName(tabName)
}
\arguments{
\item{tabName}{Tab to validate.}
}
\value{
An error if a wrong pattern is found
}
\description{
TO avoid JS issues: avoid punctuation and space
}
