% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7-inputs.R
\name{createInputLayout}
\alias{createInputLayout}
\title{Create common input layout}
\usage{
createInputLayout(..., label = NULL, style, dropdown = FALSE)
}
\description{
See \url{https://framework7.io/docs/inputs#inputs-layout}.
}
\keyword{internal}
