% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7-inputs.R
\name{createRadioOptions}
\alias{createRadioOptions}
\title{Generates a list of option for \link{f7Radio}}
\usage{
createRadioOptions(choices, selected, inputId)
}
\arguments{
\item{choices}{List of choices.}

\item{selected}{Selected value}

\item{inputId}{Radio input id.}
}
\description{
Generates a list of option for \link{f7Radio}
}
\keyword{internal}
