% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7Badge.R
\name{f7Badge}
\alias{f7Badge}
\title{Create a Framework7 badge}
\usage{
f7Badge(..., color = NULL)
}
\arguments{
\item{...}{Badge content. Avoid long text.}

\item{color}{Badge color: see here for valid colors \url{https://framework7.io/docs/badge.html}.}
}
\description{
Build a Framework7 badge
}
\examples{
if(interactive()){
 library(shiny)
 library(shinyF7)

 shiny::shinyApp(
  ui = f7Page(
    f7Init(theme = "auto"),
    title = "Badges",
    f7Badge(32, color = "blue"),
    f7Badge("Badge", color = "green")
  ),
  server = function(input, output) {}
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
