% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7Block.R
\name{f7BlockFooter}
\alias{f7BlockFooter}
\title{Framework7 block footer}
\usage{
f7BlockFooter(text = NULL)
}
\arguments{
\item{text}{Any text.}
}
\description{
\link{f7BlockFooter} creates a footer content for \link{f7Block}.
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
