% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7Block.R
\name{f7BlockTitle}
\alias{f7BlockTitle}
\title{Create a Framework7 block title}
\usage{
f7BlockTitle(title = NULL, size = NULL)
}
\arguments{
\item{title}{Block title.}

\item{size}{Block title size. NULL by default or "medium", "large".}
}
\description{
Build a Framework7 block title
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
