% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7Button.R
\name{f7Button}
\alias{f7Button}
\title{Create a Framework7 button}
\usage{
f7Button(
  inputId = NULL,
  label = NULL,
  src = NULL,
  color = NULL,
  fill = TRUE,
  outline = FALSE,
  shadow = FALSE,
  rounded = FALSE,
  size = NULL
)
}
\arguments{
\item{inputId}{The input slot that will be used to access the value.}

\item{label}{The contents of the button or link–usually a text label,
but you could also use any other HTML, like an image or \link{f7Icon}.}

\item{src}{Button link.}

\item{color}{Button color. Not compatible with outline.
See here for valid colors \url{https://framework7.io/docs/badge.html}.}

\item{fill}{Fill style. TRUE by default. Not compatible with outline}

\item{outline}{Outline style. FALSE by default. Not compatible with fill.}

\item{shadow}{Button shadow. FALSE by default. Only for material design.}

\item{rounded}{Round style. FALSE by default.}

\item{size}{Button size. NULL by default but also "large" or "small".}
}
\description{
Build a Framework7 button
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
