% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7Chips.R
\name{f7Chip}
\alias{f7Chip}
\title{Create a Framework7 chips}
\usage{
f7Chip(
  label = NULL,
  img = NULL,
  icon = NULL,
  outline = FALSE,
  status = NULL,
  icon_status = NULL,
  closable = FALSE
)
}
\arguments{
\item{label}{Chip label.}

\item{img}{Chip image, if any.}

\item{icon}{Icon, if any. IOS and Material icons available.}

\item{outline}{Whether to outline chip. FALSE by default.}

\item{status}{Chip color: see here for valid colors \url{https://framework7.io/docs/chips.html}.}

\item{icon_status}{Chip icon color: see here for valid colors \url{https://framework7.io/docs/chips.html}.}

\item{closable}{Whether to close the chip. FALSE by default.}
}
\description{
Build a Framework7 chips
}
\note{
Not ready for production color and icon isssues.
}
\examples{
if(interactive()){
 library(shiny)
 library(shinyMobile)

 shiny::shinyApp(
  ui = f7Page(
    title = "Chips",
    f7SingleLayout(
     navbar = f7Navbar(title = "f7Navbar"),
     f7Block(
     strong = TRUE,
     f7Chip(label = "Example Chip"),
     f7Chip(label = "Example Chip", outline = TRUE),
     f7Chip(label = "Example Chip", icon = f7Icon("add_round"), icon_status = "pink"),
     f7Chip(label = "Example Chip", img = "http://lorempixel.com/64/64/people/9/"),
     f7Chip(label = "Example Chip", closable = TRUE),
     f7Chip(label = "Example Chip", status = "green"),
     f7Chip(label = "Example Chip", status = "green", outline = TRUE)
    )
    )
  ),
  server = function(input, output) {}
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
