% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7-deprecated.R
\name{f7Col}
\alias{f7Col}
\title{Framework7 column container}
\usage{
f7Col(...)
}
\arguments{
\item{...}{Column content. The width is automatically handled depending
on the number of columns.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\link{f7Grid} is a replacement
}
\note{
The dark theme does not work for items embedded in a column. Use \link{f7Flex}
instead.
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
\keyword{internal}
