% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7-inputs.R
\name{f7Date}
\alias{f7Date}
\title{Create an f7 date input}
\usage{
f7Date(inputId, label, value = "", placeholder = NULL)
}
\arguments{
\item{inputId}{Date input id.}

\item{label}{Date input label.}

\item{value}{Date input value.}

\item{placeholder}{Date input placeholder.}
}
\description{
This does only work for mobiles or tablets!
}
\examples{
if(interactive()){
 library(shiny)
 library(shinyF7)

 shiny::shinyApp(
   ui = f7Page(
     title = "My app",
     f7SingleLayout(
      navbar = f7Navbar(title = "f7Date"),
      f7Date(inputId = "date", label = "Date", value = "2014-04-30"),
      verbatimTextOutput("datevalue")
     )
   ),
   server = function(input, output) {
     output$datevalue <- renderPrint({ input$date })
   }
 )
}
}
