% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7Dialog.R
\name{f7Dialog}
\alias{f7Dialog}
\title{Create a Framework7 dialog window}
\usage{
f7Dialog(text, session)
}
\arguments{
\item{text}{Dialog text.}

\item{session}{shiny session.}
}
\description{
Create a Framework7 dialog window
}
\examples{
if (interactive()) {
  library(shiny)
  library(shinyF7)
  shinyApp(
    ui = f7Page(
      title = "My App",
      f7SingleLayout(
        navbar = f7Navbar(title = "f7Dialog"),
        f7Button(inputId = "goButton", "Go!")
      )
    ),
    server = function(input, output, session) {
      shiny::observeEvent(input$goButton,{
        f7Dialog(
         text = "This is an alert dialog",
         session
        )
      })
    }
  )
}
}
