% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7Card.R
\name{f7ExpandableCard}
\alias{f7ExpandableCard}
\title{Create a Framework7 expandable card}
\usage{
f7ExpandableCard(..., id = NULL, title = NULL, subtitle = NULL,
  color = NULL)
}
\arguments{
\item{...}{Card content.}

\item{id}{Unique card id. Useful to handle multiple cards in the DOM.}

\item{title}{Card title.}

\item{subtitle}{Card subtitle.}

\item{color}{Card background color. See \url{http://framework7.io/docs/cards.html}.}
}
\description{
Build a Framework7 expandable card
}
\note{
Does not render well without \link{f7SingleLayout} or \link{f7TabLayout}.
}
\examples{
if(interactive()){
 library(shiny)
 library(shinyF7)

 shiny::shinyApp(
  ui = f7Page(
    title = "Expandable Cards",
    f7Init(theme = "auto"),
    f7SingleLayout(
      navbar = f7Navbar(
       title = "Progress",
       hairline = FALSE,
       shadow = TRUE
      ),
      f7ExpandableCard(
       id = "card1",
       title = "Expandable Card 1",
       color = "blue",
       subtitle = "Click on me pleaaaaase",
       "Framework7 - is a free and open source HTML mobile framework
       to develop hybrid mobile apps or web apps with iOS or Android
       native look and feel. It is also an indispensable prototyping apps tool
       to show working app prototype as soon as possible in case you need to."
      ),
      f7ExpandableCard(
       id = "card2",
       title = "Expandable Card 2",
       color = "green",
       "Framework7 - is a free and open source HTML mobile framework
       to develop hybrid mobile apps or web apps with iOS or Android
       native look and feel. It is also an indispensable prototyping apps tool
       to show working app prototype as soon as possible in case you need to."
      )
    )
  ),
  server = function(input, output) {}
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
