% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7Fab.R
\name{f7Fab}
\alias{f7Fab}
\title{Create a Framework7 floating action button (FAB)}
\usage{
f7Fab(inputId, label, width = NULL, ..., flag = NULL)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{label}{The contents of the button or link--usually a text label, but
you could also use any other HTML, like an image.}

\item{width}{The width of the input, e.g. \code{'400px'}, or \code{'100\%'};
see \code{\link[=validateCssUnit]{validateCssUnit()}}.}

\item{...}{Named attributes to be applied to the button or link.}

\item{flag}{Additional text displayed next to the button content. Only works
if \link{f7Fabs} position parameter is not starting with center-...}
}
\description{
Build a Framework7 floating action button (FAB)
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
