% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7Fab.R
\name{f7FabClose}
\alias{f7FabClose}
\title{Framework7 FAB close}
\usage{
f7FabClose(tag)
}
\arguments{
\item{tag}{Target tag.}
}
\description{
\link{f7FabClose} indicates that the current tag should close the \link{f7Fabs}.
}
