% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7-deprecated.R
\name{f7Flex}
\alias{f7Flex}
\title{Framework7 flex container}
\usage{
f7Flex(...)
}
\arguments{
\item{...}{Items.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\link{f7Grid} is a replacement
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
\keyword{internal}
