% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typography.R
\name{f7Float}
\alias{f7Float}
\title{Create a Framework7 float}
\usage{
f7Float(tag, side = c("left", "right"))
}
\arguments{
\item{tag}{Tag to float.}

\item{side}{Side to float: "left" or "right".}
}
\description{
Build a Framework7 float
}
\examples{
if(interactive()){
 library(shiny)
 library(shinyF7)

 shiny::shinyApp(
   ui = f7Page(
    title = "My app",
    f7Float(h1("Left"), side = "left"),
    f7Float(h1("Right"), side = "right")
   ),
   server = function(input, output) {}
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
