% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7Init.R
\name{f7Init}
\alias{f7Init}
\title{Custom initialization}
\usage{
f7Init(theme = c("ios", "md", "auto"))
}
\arguments{
\item{theme}{App theme: "ios", "md", "auto".}
}
\description{
Use inside \link{f7Page}.
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
