% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7Init.R
\name{f7Init}
\alias{f7Init}
\title{Custom initialization}
\usage{
f7Init(theme = c("ios", "md", "auto", "aurora"), filled = FALSE,
  color = NULL)
}
\arguments{
\item{theme}{App theme: "ios", "md", "auto" or "aurora".}

\item{filled}{Whether to fill the \link{f7Navbar} and \link{f7Toolbar} with
the current selected color. FALSE by default.}

\item{color}{Color theme: See \url{http://framework7.io/docs/color-themes.html}.
Expect a name like blue or red. If NULL, use the default color.}
}
\description{
Use inside \link{f7Page}. Mandatory!
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
