% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7Link.R
\name{f7Link}
\alias{f7Link}
\title{Create a Framework7 link}
\usage{
f7Link(label = NULL, icon = NULL, src = NULL, external = FALSE)
}
\arguments{
\item{label}{Link text.}

\item{icon}{Link icon, if any.}

\item{src}{Link source, url.}

\item{external}{Whether switch to an external link. FALSE by default.}
}
\description{
Build a Framework7 link
}
\examples{
if(interactive()){
 library(shiny)
 library(shinyF7)

 shiny::shinyApp(
   ui = f7Page(
    title = "Links",
    f7SingleLayout(
     navbar = f7Navbar(title = "f7Link"),
     f7Link(label = "Google", src = "https://www.google.com"),
     f7Link(label = "Google", src = "https://www.google.com", external = TRUE)
    )
   ),
   server = function(input, output) {}
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
