% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7Link.R
\name{f7Link}
\alias{f7Link}
\title{Framework7 link}
\usage{
f7Link(label, href, icon = NULL)
}
\arguments{
\item{label}{Link text.}

\item{href}{Link source, url.}

\item{icon}{Link icon, if any.}
}
\description{
Link to point toward external content.
}
\examples{
if(interactive()){
 library(shiny)
 library(shinyMobile)

 shinyApp(
   ui = f7Page(
    title = "Links",
    f7SingleLayout(
     navbar = f7Navbar(title = "f7Link"),
     f7Link(label = "Google", href = "https://www.google.com"),
     f7Link(label = "Twitter", href = "https://www.twitter.com")
    )
   ),
   server = function(input, output) {}
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
