% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7List.R
\name{f7ListIndex}
\alias{f7ListIndex}
\title{Create a Framework 7 list index}
\usage{
f7ListIndex(id, target, ..., session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{id}{Unique id.}

\item{target}{Related list element. CSS selector like .class, #id, ...}

\item{...}{Other options (see \url{https://framework7.io/docs/list-index#list-index-parameters}).}

\item{session}{Shiny session object.}
}
\description{
List index must be attached to an existing list view.
}
\note{
We disadvise to use multiple list index widget per app.
}
\examples{
if (interactive()) {
  library(shiny)
  library(shinyMobile)
  shinyApp(
    ui = f7Page(
      title = "List Index",
      f7SingleLayout(
        navbar = f7Navbar(
          title = "f7ListIndex"
        ),
        f7List(
          mode = "contacts",
          lapply(1:26, function(i) {
            f7ListGroup(
              title = LETTERS[i],
              lapply(1:26, function(j) f7ListItem(letters[j]))
            )
          })
        )
      )
    ),
    server = function(input, output, session) {
      f7ListIndex(id = "mylist", target = ".list", label = TRUE)
    }
  )
}
}
