% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7List.R
\name{f7ListIndexItem}
\alias{f7ListIndexItem}
\title{Create a Framework 7 list index item}
\usage{
f7ListIndexItem(...)
}
\arguments{
\item{...}{Item content.}
}
\description{
Create a Framework 7 list index item
}
