% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7List.R
\name{f7ListItem}
\alias{f7ListItem}
\title{Create a Framework 7 contact item}
\usage{
f7ListItem(
  ...,
  title = NULL,
  subtitle = NULL,
  header = NULL,
  footer = NULL,
  url = NULL,
  media = NULL,
  right = NULL
)
}
\arguments{
\item{...}{Item text.}

\item{title}{Item title.}

\item{subtitle}{Item subtitle.}

\item{header}{Item header. Do not use when \link{f7List} mode is not NULL.}

\item{footer}{Item footer. Do not use when \link{f7List} mode is not NULL.}

\item{url}{Item url.}

\item{media}{Expect \link{f7Icon} or \link[shiny]{img}.}

\item{right}{Right content if any.}
}
\description{
Create a Framework 7 contact item
}
