% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7Messages.R
\name{f7Message}
\alias{f7Message}
\title{Create a framework7 message}
\usage{
f7Message(
  content,
  src = NULL,
  author = NULL,
  date = NULL,
  state = c("sent", "received"),
  type = c("text", "img")
)
}
\arguments{
\item{content}{Message content. If type is ing, content must
be an url or path.}

\item{src}{Message author avatar}

\item{author}{Message author}

\item{date}{Message date}

\item{state}{Whether it is a received or sent message.}

\item{type}{Whether it is a text or image.}
}
\description{
Create a framework7 message
}
