% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7Navbar.R
\name{f7Navbar}
\alias{f7Navbar}
\title{Create a Framework7 Navbar}
\usage{
f7Navbar(subNavbar = NULL, title = NULL, subtitle = NULL,
  hairline = TRUE, shadow = TRUE, bigger = FALSE,
  left_panel = FALSE, right_panel = FALSE, hideOnScroll = FALSE)
}
\arguments{
\item{subNavbar}{\link{f7SubNavbar} slot, if any.}

\item{title}{Navbar title.}

\item{subtitle}{Navbar subtitle.}

\item{hairline}{Whether to display a thin border on the top of the navbar. TRUE by default.}

\item{shadow}{Whether to display a shadow. TRUE by default.}

\item{bigger}{Whether to display bigger title. FALSE by default.}

\item{left_panel}{Whether to enable the left panel. FALSE by default.}

\item{right_panel}{Whether to enable the right panel. FALSE by default.}

\item{hideOnScroll}{Whether to hide the navbar on scroll. FALSE by default.}
}
\description{
Build a Framework7 Navbar
}
\note{
hideOnScroll does not work yet.
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
