% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7Navbar.R
\name{f7Navbar}
\alias{f7Navbar}
\title{Create a Framework7 Navbar}
\usage{
f7Navbar(..., subNavbar = NULL, title = NULL, subtitle = NULL,
  hairline = TRUE, shadow = TRUE, bigger = FALSE,
  transparent = FALSE, left_panel = FALSE, right_panel = FALSE)
}
\arguments{
\item{...}{Slot for \link{f7SearchbarTrigger}. Not compatible with \link{f7Panel}.}

\item{subNavbar}{\link{f7SubNavbar} slot, if any.}

\item{title}{Navbar title.}

\item{subtitle}{Navbar subtitle. Not compatible with bigger.}

\item{hairline}{Whether to display a thin border on the top of the navbar. TRUE by default.}

\item{shadow}{Whether to display a shadow. TRUE by default.}

\item{bigger}{Whether to display bigger title. FALSE by default. Not
compatible with subtitle.}

\item{transparent}{Whether the navbar should be transparent. FALSE by default.
Only works if bigger is TRUE.}

\item{left_panel}{Whether to enable the left panel. FALSE by default.}

\item{right_panel}{Whether to enable the right panel. FALSE by default.}
}
\description{
Build a Framework7 Navbar
}
\note{
Currently, bigger parameters does mess with the CSS.
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
