% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7Page.R
\name{f7Page}
\alias{f7Page}
\title{Create a Framework7 page}
\usage{
f7Page(..., init = f7Init(theme = "auto"), title = NULL,
  dark_mode = FALSE, color = NULL)
}
\arguments{
\item{...}{Any element.}

\item{init}{App configuration. See \link{f7Init}.}

\item{title}{Page title.}

\item{dark_mode}{Whether to enable the dark mode. FALSE by default.}

\item{color}{Color theme: See \url{http://framework7.io/docs/color-themes.html}.}
}
\description{
Build a Framework7 page
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
