% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7Page.R
\name{f7Page}
\alias{f7Page}
\title{Create a Framework7 page}
\usage{
f7Page(..., title = NULL, dark_mode = FALSE, color = NULL)
}
\arguments{
\item{...}{Any element.}

\item{title}{Page title.}

\item{dark_mode}{Whether to enable the dark mode. FALSE by default.}

\item{color}{Color theme: See \url{http://framework7.io/docs/color-themes.html}.}
}
\description{
Build a Framework7 page
}
\examples{
if(interactive()){
 library(shiny)
 library(shinyF7)

 shiny::shinyApp(
   ui = f7Page(title = "My app"),
   server = function(input, output) {}
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
