% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7Page.R
\name{f7Page}
\alias{f7Page}
\title{Create a Framework7 page}
\usage{
f7Page(
  ...,
  init = f7Init(skin = "auto", theme = "light"),
  title = NULL,
  preloader = FALSE,
  loading_duration = 3
)
}
\arguments{
\item{...}{Slot for shinyMobile skeleton elements: \link{f7Appbar}, \link{f7SingleLayout},
\link{f7TabLayout}, \link{f7SplitLayout}.}

\item{init}{App configuration. See \link{f7Init}.}

\item{title}{Page title.}

\item{preloader}{Whether to display a preloader before the app starts.
FALSE by default.}

\item{loading_duration}{Preloader duration.}
}
\description{
Build a Framework7 page
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
