% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7Panel.R
\name{f7Panel}
\alias{f7Panel}
\title{Create a Framework7 panel}
\usage{
f7Panel(..., side = c("left", "right"), theme = c("dark", "light"))
}
\arguments{
\item{...}{Panel content.}

\item{side}{Panel side: "left" or "right".}

\item{theme}{Panel background color: "dark" or "light".}
}
\description{
Build a Framework7 panel
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
