% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7Panel.R
\name{f7Panel}
\alias{f7Panel}
\title{Create a Framework7 panel}
\usage{
f7Panel(..., title = NULL, side = c("left", "right"),
  theme = c("dark", "light"), style = c("reveal", "cover"),
  resizable = FALSE)
}
\arguments{
\item{...}{Panel content.}

\item{title}{Panel title.}

\item{side}{Panel side: "left" or "right".}

\item{theme}{Panel background color: "dark" or "light".}

\item{style}{Whether the panel should behave when opened: "cover" or "reveal".}

\item{resizable}{Whether to enable panel resize. FALSE by default.}
}
\description{
Build a Framework7 panel
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
