% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7Panel.R
\name{f7PanelItem}
\alias{f7PanelItem}
\title{Create a Framework7 sidebar menu item}
\usage{
f7PanelItem(id)
}
\arguments{
\item{id}{Item id.}
}
\description{
Build a Framework7 sidebar menu item
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
