% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7Progress.R
\name{f7ProgressInf}
\alias{f7ProgressInf}
\title{Create a Framework7 infinite progress bar}
\usage{
f7ProgressInf(color = NULL)
}
\arguments{
\item{color}{Progress color. See \url{http://framework7.io/docs/progressbar.html}.}
}
\description{
Build a Framework7 infinite progress bar
}
\note{
Buggy display when status is not NULL.
}
\examples{
if(interactive()){
 library(shiny)
 library(shinyMobile)

 shiny::shinyApp(
   ui = f7Page(
    title = "Progress Infinite",
    f7SingleLayout(
     navbar = f7Navbar(title = "f7ProgressInf"),
     f7Block(
     f7ProgressInf(),
     f7ProgressInf(color = "yellow")
    )
    )
   ),
   server = function(input, output) {}
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
