% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7Searchbar.R
\name{f7SearchbarTrigger}
\alias{f7SearchbarTrigger}
\title{Create a Framework 7 searchbar trigger}
\usage{
f7SearchbarTrigger(targetId)
}
\arguments{
\item{targetId}{Id of the \link{f7Searchbar}}
}
\description{
Create a Framework 7 searchbar trigger
}
\examples{
if (interactive()) {

}
}
