% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7Navbar.R
\name{f7ShowNavbar}
\alias{f7ShowNavbar}
\title{Show a Framework7 navbar}
\usage{
f7ShowNavbar(animate = TRUE, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{animate}{Whether it should be hidden with animation or not. By default is TRUE.}

\item{session}{Shiny session object.}
}
\description{
\link{f7ShowNavbar} shows a \link{f7Navbar} component from the
server. Deprecated, use \link{updateF7Navbar} instead.
}
