% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7Navbar.R
\name{f7ShowNavbar}
\alias{f7ShowNavbar}
\title{Show a framework7 navbar}
\usage{
f7ShowNavbar(animate = TRUE, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{animate}{Whether it should be hidden with animation or not. By default is TRUE.}

\item{session}{Shiny session object.}
}
\description{
Show a framework7 navbar
}
\examples{
if (interactive()) {
 library(shiny)
 library(shinyMobile)

 shinyApp(
    ui = f7Page(
       title = "Show navbar",
       f7SingleLayout(
          navbar = f7Navbar("Hide/Show navbar"),
          f7Segment(
             f7Button(inputId = "hide", "Hide navbar", color = "red"),
             f7Button(inputId = "show", "Show navbar", color = "green"),
          )
       )
    ),
    server = function(input, output, session) {

       observeEvent(input$hide, {
          f7HideNavbar()
       })

       observeEvent(input$show, {
          f7ShowNavbar()
       })
    }
 )
}
}
