% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7Page.R
\name{f7SidebarLayout}
\alias{f7SidebarLayout}
\title{Create a Framework7 page with split layout}
\usage{
f7SidebarLayout(sidebarPanel, mainPanel)
}
\arguments{
\item{sidebarPanel}{Slot for \link{f7Sidebar}.}

\item{mainPanel}{Slot for \link{f7mainPanel}.}
}
\description{
Build a Framework7 page with split layout
}
\examples{
if(interactive()){
 library(shiny)
 library(shinyF7)

 shiny::shinyApp(
  ui = f7Page(
    setTheme("ios"),
    title = "Sidebar Layout",
    f7SidebarLayout(
      # sidebar content
      sidebarPanel = f7Sidebar(
        title = "Sidebar",
        side = "left",
        theme = "light",
        sliderInput("obs", "Number of observations", 0, 1000, 500)
      ),
      # main content
      mainPanel = f7mainPanel(
        navbar = f7Navbar(
          title = "Sidebar Layout",
          hairline = FALSE,
          shadow = TRUE
        ),
        toolbar = f7Toolbar(
          f7Link(label = "Link 1", src = "https://www.google.com"),
          f7Link(label = "Link 2", src = "https://www.google.com", external = TRUE)
        ),
        f7Shadow(
          intensity = 10,
          hover = TRUE,
          f7Card(
            title = "Card header",
            plotOutput("distPlot"),
            footer = tagList(
              f7Button(color = "blue", "My button", src = "https://www.google.com"),
              f7Badge("Badge", color = "green")
            )
          )
        )
      )
    )
  ),
  server = function(input, output) {
    output$distPlot <- renderPlot({
      dist <- rnorm(input$obs)
      hist(dist)
    })
  }
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
