% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7-tabs.R
\name{f7Tab}
\alias{f7Tab}
\title{Create a Framework7 tab}
\usage{
f7Tab(..., id, active = FALSE)
}
\arguments{
\item{...}{Any UI element.}

\item{id}{Unique tab id.}

\item{active}{Whether the tab is active at start.}
}
\description{
Build a Framework7 tab
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
