% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7-tabs.R
\name{f7Tab}
\alias{f7Tab}
\title{Create a Framework7 tab item}
\usage{
f7Tab(..., tabName, icon = NULL, active = FALSE)
}
\arguments{
\item{...}{Item content.}

\item{tabName}{Item id. Must be unique.}

\item{icon}{Item icon.}

\item{active}{Whether the tab is active at start. Do not select multiple tabs, only
the first one will be set to active}
}
\description{
Build a Framework7 tab item
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
