% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7-tables.R
\name{f7Table}
\alias{f7Table}
\title{Framework7 table}
\usage{
f7Table(data, colnames = NULL, card = FALSE)
}
\arguments{
\item{data}{A data.frame.}

\item{colnames}{Column names to use, if \code{NULL} uses \code{data} column names.}

\item{card}{Whether to use as card.}
}
\description{
Creates a table container.
}
\examples{
library(shiny)
library(shinyMobile)

app <- shinyApp(
  ui = f7Page(
    title = "My app",
    f7SingleLayout(
      navbar = f7Navbar(
        title = "f7Table"
      ),
      uiOutput("table")
    )
  ),
  server = function(input, output) {
    output$table <- renderUI({
      f7Table(mtcars)
    })
  }
)

if (interactive() || identical(Sys.getenv("TESTTHAT"), "true")) app
}
