% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7-tabs.R
\name{f7Tabs}
\alias{f7Tabs}
\title{Create a Framework7 tabs}
\usage{
f7Tabs(..., id = NULL, swipeable = FALSE, animated = TRUE)
}
\arguments{
\item{...}{Slot for \link{f7Tab}.}

\item{id}{Optional to get the id of the currently selected \link{f7Tab}.}

\item{swipeable}{Whether to allow finger swip. FALSE by default. Only for touch-screens.
Not compatible with animated.}

\item{animated}{Whether to show transition between tabs. TRUE by default.
Not compatible with swipeable.}
}
\description{
Build a Framework7 tabs
}
\note{
Animated does not work when set to FALSE and swipeable is FALSE.
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
