% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7Toolbar.R
\name{f7Toolbar}
\alias{f7Toolbar}
\title{Create a Framework7 Toolbar}
\usage{
f7Toolbar(...)
}
\arguments{
\item{...}{Slot for \link{f7Link} or any other element.}
}
\description{
Build a Framework7 Toolbar
}
\note{
Does not render properly on iOS.
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
