% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7-toolbars.R
\name{f7Toolbar}
\alias{f7Toolbar}
\title{Create a Framework7 toolbar/tabbar}
\usage{
f7Toolbar(..., bottom = FALSE, hairline = TRUE, shadow = TRUE,
  tabs = FALSE, labels = FALSE, scrollable = FALSE)
}
\arguments{
\item{...}{Slot for f7ToolbarItem.}

\item{bottom}{Toolbar position. FALSE by default, the toolbar is displayed
below the navbar. If TRUE, it is displayed at the bottom of the page.}

\item{hairline}{Whether to display a thin border on the top of the toolbar. TRUE by default.}

\item{shadow}{Whether to display a shadow. TRUE by default.}

\item{tabs}{Whether to use the toolbar with tabs. See f7TabList and f7Tab. FALSE
by default.}

\item{labels}{Whether to use icons instead of text. Only use when tabbar is TRUE.}

\item{scrollable}{Whether to allow scrolling. FALSE by default.}
}
\description{
Build a Framework7 toolbar/tabbar
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
