% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7Toolbar.R
\name{f7Toolbar}
\alias{f7Toolbar}
\title{Framework7 Toolbar}
\usage{
f7Toolbar(
  ...,
  position = c("top", "bottom"),
  hairline = TRUE,
  shadow = TRUE,
  icons = FALSE,
  scrollable = FALSE
)
}
\arguments{
\item{...}{Slot for \link{f7Link} or any other element.}

\item{position}{Tabs position: "top" or "bottom".}

\item{hairline}{Whether to display a thin border on the top of the toolbar. TRUE by default.}

\item{shadow}{Whether to display a shadow. TRUE by default.}

\item{icons}{Whether to use icons instead of text. Either ios or md icons.}

\item{scrollable}{Whether to allow scrolling. FALSE by default.}
}
\description{
\code{f7Toolbar} is a layout element located at the bottom or top. It is
internally used by \link{f7Tabs}.
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
