% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7Tooltip.R
\name{f7Tooltip}
\alias{f7Tooltip}
\title{Create a Framework7 tooltip}
\usage{
f7Tooltip(tag, text)
}
\arguments{
\item{tag}{Tooltip target.}

\item{text}{Tooltip content.}
}
\description{
This uses the auto init framework 7 tooltip
}
\examples{
if (interactive()) {
 library(shiny)
 library(shinyMobile)
 shinyApp(
   ui = f7Page(
     color = "pink",
     title = "My app",
     f7SingleLayout(
       navbar = f7Navbar(title = "f7Tooltip"),
       f7Tooltip(
         f7Badge("Hover on me", color = "pink"),
         text = "A tooltip!"
       )
     )
   ),
   server = function(input, output, session) {
   }
 )
}
}
