% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7-update-inputs.R
\name{updateF7DatePicker}
\alias{updateF7DatePicker}
\title{Change the value of a date picker input on the client}
\usage{
updateF7DatePicker(
  session,
  inputId,
  value = NULL,
  choices = NULL,
  rotateEffect = NULL,
  openIn = NULL,
  scrollToInput = NULL,
  closeByOutsideClick = NULL,
  toolbar = NULL,
  toolbarCloseText = NULL,
  sheetSwipeToClose = NULL
)
}
\arguments{
\item{session}{The session object passed to function given to the server.}

\item{inputId}{The id of the input object.}

\item{value}{Picker initial value, if any.}

\item{choices}{New picker choices.}

\item{rotateEffect}{Enables 3D rotate effect. Default to TRUE.}

\item{openIn}{Can be auto, popover (to open picker in popover), sheet (to open in sheet modal).
In case of auto will open in sheet modal on small screens and in popover on large screens. Default
to auto.}

\item{scrollToInput}{Scroll viewport (page-content) to input when picker opened. Default
to FALSE.}

\item{closeByOutsideClick}{If enabled, picker will be closed by clicking outside of picker or related input element.
Default to TRUE.}

\item{toolbar}{Enables picker toolbar. Default to TRUE.}

\item{toolbarCloseText}{Text for Done/Close toolbar button.}

\item{sheetSwipeToClose}{Enables ability to close Picker sheet with swipe. Default to FALSE.}
}
\description{
Change the value of a date picker input on the client
}
\examples{
if (interactive()) {
 library(shiny)
 library(shinyMobile)

 shinyApp(
  ui = f7Page(
    title = "My app",
    f7SingleLayout(
      navbar = f7Navbar(title = "Update date picker"),
      f7Card(
        f7Button(inputId = "update", label = "Update date picker"),
        f7DatePicker(
          inputId = "mypicker",
          label = "Choose a date",
          value = "2019-08-24",
          openIn = "auto",
          direction = "horizontal"
        ),
        verbatimTextOutput("pickerval"),
        br(),
        f7Button(inputId = "removeToolbar", label = "Remove date picker toolbar", color = "red")
      )
    )
  ),
  server = function(input, output, session) {

    output$pickerval <- renderText(input$mypicker)

    observeEvent(input$update, {
      updateF7DatePicker(
        session,
        inputId = "mypicker",
        value = "2019-08-30",
        openIn = "sheet",
        toolbarCloseText = "Prout",
        sheetSwipeToClose = TRUE
      )
    })

    observeEvent(input$removeToolbar, {
      updateF7DatePicker(
        session,
        inputId = "mypicker",
        openIn = "sheet",
        toolbar = FALSE
      )
    })

  }
 )
}
}
