% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7-tabs.R
\name{updateF7Tabs}
\alias{updateF7Tabs}
\title{Update a Framework 7 tabsetPanel}
\usage{
updateF7Tabs(id, selected = NULL, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{id}{Id of the \link{f7Tabs} to update.}

\item{selected}{Newly selected tab.}

\item{session}{Shiny session object.}
}
\description{
Update \link{f7Tabs}.
}
\seealso{
\link{f7Tabs}
}
